Add application icon and integrate into project
Added `iconBase64.ico` as the application icon by including it in the project file (`ConverterBase64.csproj`) and specifying it as the `ApplicationIcon`. Updated `FrmPrincipal.Designer.cs` to set the form's `Icon` property using a resource object.

Embedded the `iconBase64.ico` file as a resource in `FrmPrincipal.resx` using a base64-encoded byte array. Added a reference to the `System.Drawing` assembly in `FrmPrincipal.resx` to support the use of the icon.